requirements/dev.txt
-r base.txt
pytest>=8.0.0
pytest-django>=4.9.0
ruff>=0.6.0
